Implement concurrency managed workqueue
Use the concurrency managed workqueue api (cmwq) to create a dedicated
workqueue instead of using the default system_wq. The list management
functions from the kernel api are used to manage metadata of workers.
As a result, it is guaranteed that all resource used by workers will be
release when removing the khttpd module.